Run integration tests with all serializers _only_ for Filesystem backend; for all other backends, only test with default serializer
