+++ pyproject.toml
@@ -137,6 +137,7 @@ omit = [
 exclude_lines = [
     'pragma: no cover',
     'if TYPE_CHECKING:',
+    'except ImportError:',
 ]
 
 [tool.isort]
